pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-logseq"
version = "0.1.0"
description = "LangChain helpers for working with Logseq documents"
authors = [{ name="DL", email="v49t9zpqd@mozmail.com" }]
license = { text = "MIT" }
dependencies = [
    "langchain>=0.3.24,<0.4.0",
    "pydantic>=2.11,<3.0",

    "pgvector-template @ git+https://github.com/DavidLiuGit/PGVector-Template.git",
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
scripts = [
  "python-dotenv",
  "psycopg2-binary>=2.9.0",
]
test = [
  "python-dotenv",
  "boto3",
  "boto3-stubs",
  "langchain-aws",
]

[tool.setuptools.packages.find]
where = ["."]  # Look for packages in the root directory
include = ["langchain_logseq*"]  # Include all packages starting with langchain_logseq


[tool.black]
line-length = 120
target-version = ["py311", "py312", "py313"]
